_eel/requirements.txt
eel
pyyaml
markdown
pygments
